requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3

# Dashboard
streamlit==1.28.1
plotly==5.18.0
pandas==2.1.1
numpy==1.26.1

# NLP and LLM
langchain==0.1.0
langchain-community==0.0.10
openai==1.3.0
tiktoken==0.5.1

# Networking
aiohttp==3.8.6
requests==2.31.0

# Data Parsing and Manipulation
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2

# Graph Visualization
networkx==3.2.1

# Utilities
python-dotenv==1.0.0
pytest==7.4.3 